avs-nwtest: implement via jsonnet/kubecfg (WIP)
This implements deployment of the AVS Network Test Tool
(avs-nwtesttool).

To make it a bit more interesting, we deploy it using kubecfg/jsonnet
instead of using Helm.

For that, we pull in `kube.libsonnet` from bitnami, and write a small
framework that is composed of:
 - wire deployments/environments
   (we call them deployment at the top-level configs, but environments
   deepr within jsonnet machinery to not confuse them with Kubernetes
   Deployments)
 - wire components
   (eg. microservices composed usually of a deployment+service that is
   exposed via nginz - this is a paradigm that can be extended in the
   future)

We not only export kubernetes-consumable objects, but also nginz routes
as an example.

This is all currently documented in a small README.md, which will be
turned into real documentation once this PR stabilizes.