*: unify DNS names for demo deployments
This change unifies the list of domains that users should set up to the
following list of canonical prefixes:

 - nginz-https.<domain>
 - nginx-ssl.<domain>
 - webapp.<domain>
 - s3.<domain>
 - teams.<domain> (optional)
 - account.<domain> (optional)

Naturally, following this scheme is not a hard requirement, and users can
change these as they please. However, keeping them in sync allows for
users to do a blind search/replace from `example.com` to their domain,
without having to immediately understand the roles of different
subdomains.

Optimally we would have all of these defined in a single source of truth
file that would be sourced from all Helm charts, but I'm not sure this
is something that we can easily do.