Add TF module for brig to provide prekey locking, an event queue and (optionally) email sending services (#244)
Prekey and queue for internal events are required by brig, whereas the services for emailing is not needed and could be provided in a separate way (see chart value `brig.config.useSES`). 

But since Brig can only take one set of SQS credentials, splitting it into two modules would have required to facilitate some sort of data interface. To simplify things for now, this was approach was dropped in favour of a feature flag to enable mailing setup while keeping all the code in one TF module.

* adds DNS records for for sending emails
* hard-code the amount of dkim records (3) which are based on the amount of dkim tokens