Provide AWS_REGION variable to cargohold (#314)
* provide AWS_REGION variable to cargohold

Now with cargohold using V4 signatures
(https://github.com/wireapp/wire-server/pull/1157),
the region is part of the Authorization header and needs to configured correctly.
We already do that for the other services, so I just copied the code from there.

* provide AWS env vars to cargohold integration

We do this for all the other servers, too, but I'm actually not sure
why. The integration tests themselves don't call AWS.

We should probably just remove it (+ from the other services).