[tf-module:push-notifications] Dedup logic and SNS platform app resources
This changeset unifies both platform resources. This is possible because
TF - for once - behaves as expected and drops a resource attribute that
is set to 'null'.
This also helps to deduplicate the logic that is used to massage the app
lists inputs to cope with 'for_each'.

Allow the list of apps to not being set. Because this is now allowed, the
overall existence of any application is used as indicator to figure out
whether the IAM policy has to be created or not.