[tf-module:push-notifications] Allow to define multiple apps per client platform (#347)
Commit (1) Allow to defined multiple apps per client platform

NOTE: This is a breaking change and requires Terraform
      state migration ('terraform state mv')

Due to the fact that multiple different clients (app/build ID) can be connected to
the same backend and at the same time desire to receive push notifications, it should
be possible to register those apps to the same events/notifications queue.


Commit (2) Support multiple platforms for one app


Commit (3) Dedup logic and SNS platform app resources

This changeset unifies both platform resources. This is possible because
TF - for once - behaves as expected and drops a resource attribute that
is set to 'null'.
This also helps to deduplicate the logic that is used to massage the app
lists inputs to cope with 'for_each'.

Allow the list of apps to not being set. Because this is now allowed, the
overall existence of any application is used as indicator to figure out
whether the IAM policy has to be created or not.