demo-smtp: Don't use `latest` tag
The docker hub image contains a single, mutable `latest` tag.

Having a `latest` tag in a image in k8s causes it to have a
imagePullPolicy: always, which fails in offline setups.

Instead of this, just pin our helm chart to a known-good image by
sha256 (and remove our homegrown image:tag separation)

It seems upstream has moved to use github's container registry, however,
as this one *always* requires authentication, we don't use it for now.