charts/fake-aws-s3: disable web ui again like before
This causes helm install to print:

coalesce.go:160: warning: skipped value for environment: Not a table.
coalesce.go:160: warning: skipped value for environment: Not a table.
coalesce.go:160: warning: skipped value for environment: Not a table

but that seems to be a bug in helm and we can ignore it.

Something about the way helm handles child-charts (which minio is in
this setup)