charts/sftd: introduce
Adds a new helm chart to deploy sftd as a single replica.

This is suitable for small installations with <50 users.
In the future we will support running SFTD with multiple replicas to
support scale-out of conference calling infrastructure.

The chart runs sftd in the hostNetwork:true such that the SFT
can allocate WebRTC mediastreams on its public IP so people can connect
and call.

This is not ideal, but it's the best workaround I could come up with.

We could solve this by getting
https://github.com/kubernetes/kubernetes/issues/23864 fixed but I don't
have a lot of hope as it has been open for quite a few years.

Other way to fix this is to not use IPv4 but Ipv6 in the cluster. Then
every pod has a proper public ip address and non of this is an issue at
all to begin with. But given IPv6 adoption isn't particularly high,
and because I don't think our TURN servers could bridge the IPv4-IPv6
gap I do not think it's a realistic option for now.