experiment: Highly available SFTD in kubernetes
This implements the proposal in https://wearezeta.atlassian.net/wiki/spaces/ENGINEERIN/pages/150930455/SFT+Load+balancing

Not packed up as a helm chart, but uses kubectl's built-in `kustomize`
functionality for making configs configurable.

Happy to convert this back to helm if we find this better for external
consumption but I wanted to try out this kustomize thing.