charts/sftd: Don't set the advertise media address explicitly at all
The code that iterates the default route and binds to it is already
implemented in SFTD proper

From discussion with the Calling team:

SO, the  binding address will be the address of the default network
interface, i.e. whatever interface is used for the default gateway, and
has UP status.  essentially it uses either the route list, if the OS
supports it, and chooses the interface that is used for the default
gateway route, OR if this is not available, then what getifaddrs returns
as the first interface that has the status "UP"

We can reintroduce the -A flag later to handle the scenario where
sftd is deployed in a 1:1 NAT scenario (Like in GCP)