ansible: Use git submodules instead of ansible-galaxy and bump kubespray
This bumps kubespray to https://github.com/kubernetes-sigs/kubespray/releases/tag/v2.14.2

This moves us to kubernetes v1.18.10 as the new default.

We do not have a direct upgrade path from our previous kubespray
checkout to this one. Instead one should spin up a new Kubernetes
cluster and decomission the old one.

Our charts are currently also not compatible with v1.14 (the previous
version that we shipped) so doing this update is already sort of
mandatory.