kubespray: 2.14 -> 2.15
This is 2.15 with https://github.com/kubernetes-sigs/kubespray/pull/7247
cherry-picked on top; as we need it for offline deploy.

It will allow us to point the docker install playbook to an alternative
package repository.

Though 2.14 we never had in a release, I will still provide some upgrade
instructions:

To upgrade an existing cluster follow the instructions here:
https://kubespray.io/#/docs/upgrades?id=graceful-upgrade

```
$ cd ansible
$ ansible-playbook ./roles-external/kubespray/upgrade-cluster.yml -i ./your-inventory
```

This should guide you through the upgrade process step by step.