[tf:env+k8s-module] Restructure how a k8s cluster is defined
THis change set introduces the concept of cluster groups, which allows to define
control plane and nodes separately. In the future it can be adapted for any sort
of services, where a service might contain of one or more groups. And, it still
allows to define a single-machine k8s installation.

Features:
* allow to define a volume per group
* all k8s machines are now sharing a private network

Fixes:
* move from 'count' to 'for_each' in order to gain more control over which machine
  is created or destroyed