[tf:k8s+dns-modules+env] Separate SPF record IPs from LB IPs
THe LB usually does not send emails, thus using its IP(s) in an SPF record
does not make sense. So, we need provide the node IP(s), because the SMTP
server (deployed via Helm) will on one them.
This change set makes it possible to define a list of IPv4 addresses
specifically for the SPF record. If list is empty, no SPF record is being
created.

NOTE: it would have been nicer to generate the node IPs list similar to how
it's done in kubernetes.inventory.tf, but Terraform had other plans. At the
moment, I guess it's a bug that Tf wont see the indirect dependency. Otherwise
it would not explain why it works having the logic moved into the k8s module.