terraform: Block all traffic but DNS and NTP
This allows us to simulate an offline environment where NTP is still
available.

Cassandra needs servers to be relatively in sync. However how that is
done is out of scope for us at the moment so we assume some external NTP
server.

In the future we could perhaps set up our own NTP servers in offline
as what we care about is only _relative_ clock difference; not absolute