Offline ansible (#405)
* bump elasticsearch role

Version didn't work with ansible 2.9

* ansible: Restructure example inventory files to follow ansible documented directory structure

This is needed because the documented way to customize kubespray is
through group_vars. And this allows people to do that.

We use this for an example inventory file for offline

* bin/offline: Remove

Scripts for making a collectio nof offline charts will be moved to
wire-server in a follow-up PR; so having this code here does not make
sense anymore

* mirror apt repositories

Adds scripts that mirrors the required debian packages for kubespray
and all our other playbooks

* offline:  Build and upload docker image with offline environment

* Download all the requirements for offline kubespray and all other
ansible playbooks

Next step is pointing ansible to the offline artifact

* Offline kubespray and ansible

Set up variables such that all downloaded sources are fetched from
the assets tarball. This requires an assethost to those those assets

An example inventory file is included to showcase an offline setup

* Set turn secret to bogus placeholder value

This is just so that `helm template` succeeds with the example
values.yaml :)

* Mirror helm charts and their container dependencies

* rename default inventory file to 99-static

This means it gets picked up if you specify an inventory as a
_directory_ instead of as a file

* Some fixes

* Fix offline/ci.sh script

helm containers had wrong directory structure; also tarball wouldn't
build

* Add note to inventory about default ip address

* cassandra: Remove hostname role from cassandra

No need to override hostnames (as far as I know)

* cassandra: Do not install ntpd in offline

We don't need it.

I actually question if we need it at all?

we're not using the ntpd server functionality; only the client
functionality, but Ubuntu 18.04 ships with an NTP client by default
https://ubuntu.com/server/docs/network-ntp

But this is something for another time. Lets just conditionally disable
it

* cassandra: Use JRE; not JDK

We do not need the java compiler to run cassandra (I hope?)

* cassandra: Force AWS autodetection off

* Copy example values into offline artifact

* Enable team-settings and account-pages

* Add secrets script to generate secrets for offline

This needs to be re-integrated with secrets.sh at some point, but I
couldn't bother with the rebase for now.

* restund: Remove hostname and ansible-role-ntp roles

We have no need to modify the hostname for restund servers or to have
ntpd installed

* restund: Disable TLS turn for now

* restund: remove vars_prompt for restund_zrest_secret

This value is auto-generated now and known

* Get rid of intermediate assets directory

This allows you to run `offline/ci.sh` and then the root directory is
ready for offline deploy. Useful for interactive development

* Adjust github pipeline

Compression of artifact has been moved into offline/ci.sh script

* Add docker alias for offline environment

People should run `source ./bin/offline-env.sh`

Will add this to the docs

* move offline-cluster to bin folder

* Add wrapper script for easy remote call of install instructions

Will be used in CD

* Add offline-helm script for Continious delivery

We'll use this to test out if all the offline helm artifacts indeed
install

* Fix helm_external playbook

It didn't pick up the network interfaces unless they were in [all:vars]
which was a bit akward.

* Generate secrets and deploy helm charts

* Add continous delivery for testing offline

* Use tarball instead of gzipped tarball

Tradeoff between bandwidth and compression. And we're compressing the
entire assets tarball at the end anyway

* Add sftd to offline package

* ci.sh: Remove inbetween artifacts

Github Actions otherwise doesn't have enough disk space!

We keep the possibility for incrementality locally; as that's useful

* terraform: Block all traffic but DNS and NTP

This allows us to simulate an offline environment where NTP is still
available.

Cassandra needs servers to be relatively in sync. However how that is
done is out of scope for us at the moment so we assume some external NTP
server.

In the future we could perhaps set up our own NTP servers in offline
as what we care about is only _relative_ clock difference; not absolute

* USe ssh-agent so that we do not have to put private key in cloud-init
config

* Some fixes to the offline deploy scripts

* Fix offline-env alias

* Mark terraform outputs sensitive

* Disable CD for now

Github Actions is too slow. We run this manually on Hetzner for now.

* Add docs for offline deploy

* Update .gitignore

Co-authored-by: Lucendio <gregor.jahn@wire.com>

* Add a note about the recently introduced firewall hardening to the docs

...and how to mitigate that

* ci.si: Pin wire-server version

In the future we probably wanna move the helm chart scripts to the
wire-server repo and automatically bundle an offline deploy of the helm
charts for each wire-server release

* HACK: free op some disk space in github actions

* Update restund pin

We updated the submodule; so new container

Co-authored-by: Lucendio <gregor.jahn@wire.com>

* Set LOCALHOST_PYTHON in nix-based container

* Re-enable CD; but with cleanup handled by Github

The cleanup action will always be run; even if the job is cancelled.
This makes sure there are no dangling resources.

However we should probably just use remote state instead; to make sure
to handle all edge cases.

* Make github actions not idle the connection

* Update changlog and add support for pushing tagged releases

Also upload docker container for people who do not use airgap install

* [skip ci] Fix inventory path

* Disable pretty tags for now. There's a bug i don't want to debug at this
point

We can do the release tagging bits in a follow-up PR.

* Fix container upload

Co-authored-by: Jun Matsushita <jun@iilab.org>
Co-authored-by: Lucendio <gregor.jahn@wire.com>