[tf-module:push-notifications]: make success_feedback_sample_rate configurable (#502)
This allows setting this value to arbitrary values. Resources manually
imported with `terraform import` might have this set to a value. Without
a possibility to set this, subsequent `terraform apply` runs will
complain about not being able to remove this field.

Set this to the default value, instead of trying to fix this bug in the
aws provider.