Properly kill the backgrounded liveness loop in offline-deploy.sh
If the liveness loop in the background is only killed when the deploy script
runs normally, then if the script dies on one of the intermediate commands the
liveness process won't also be terminated.

This script switches to killing the loop process in an exit handler, so that
the kill is always executed when the script exits, to prevent stale processes
from hanging around after the script.