nix: make python versions consistent.
The Nix overlay was changed to pull in python3 (with an unqualified minor
version), which is currently an alias for Python 3.9. However, the list of
bundled Python packages was not updated similarly, which resulted in a 3.9
interpreter being bundled with a series of 3.8 modules. This breaks when
attempting to use tools which rely on these packages, such as using boto for
uploading files to S3.

Co-Authored-By: Akshay Mankar <akshay@wire.com>