nix/overlay.nix: describe how we use this GPG key
This key isn't a secret (it's built and uploaded to the binary cache
after all ;-) ) It's created out of the necessity that apt wants to
verify against a key.

It's set to expire 2y after its creation,
or whenever this derivation is built again without having the result in the binary cache.

The public part of the key is shipped with the offline bundle
($aptly_root/public/gpg).